fix(services-internal): bound conversions and simplify conditions
Implements ADR-006 (5)

+++ vca/services/internal/status/lists/record.go
@@ -119,7 +119,7 @@ func NewRecord(id string, kind Kind, purpose Purpose, bitsPerEntry, size int, is
 	}
 	allocated := make([]byte, (size+7)/8)
 	for i := size; i < len(allocated)*8; i++ {
-		allocated[i/8] |= 1 << (7 - uint(i%8))
+		allocated[i/8] |= 1 << (7 - i%8)
 	}
 	return Record{
 		ID: id, Kind: kind, Purpose: purpose, Bits: bitsPerEntry, Size: size,
@@ -135,7 +135,7 @@ func (r Record) IsAllocated(i int) bool {
 	if i < 0 || i >= r.Size {
 		return false
 	}
-	return r.Allocated[i/8]&(1<<(7-uint(i%8))) != 0
+	return r.Allocated[i/8]&(1<<(7-i%8)) != 0
 }
 
 // Allocate picks one free index at random and marks it allocated
@@ -161,7 +161,7 @@ func (r *Record) Allocate(rnd io.Reader) (int, error) {
 			continue
 		}
 		for bit := 0; bit < 8; bit++ {
-			mask := byte(1 << (7 - uint(bit)))
+			mask := byte(1 << (7 - bit))
 			if b&mask != 0 {
 				continue
 			}
